+++ src/components/screen/screen-manager.jsx
@@ -39,12 +39,12 @@ function ScreenManager({
   const { t } = useTranslation("common", { keyPrefix: "screen-manager" });
   const navigate = useNavigate();
   const [orientationOptions] = useState([
-    { title: "Vertikal", id: "vertical" },
-    { title: "Horisontal", id: "horizontal" },
+    { title: "Vertikal", "@id": "vertical" },
+    { title: "Horisontal", "@id": "horizontal" },
   ]);
   const [resolutionOptions] = useState([
-    { title: "4K", id: "4K" },
-    { title: "HD", id: "HD" },
+    { title: "4K", "@id": "4K" },
+    { title: "HD", "@id": "HD" },
   ]);
   const headerText =
     saveMethod === "PUT" ? t("edit-screen-header") : t("create-screen-header");